Final updates, strategy changes

+++ Application/DTOs/AlcoholDTO.cs
@@ -8,7 +8,7 @@ namespace DTOs
 {
     public class AlcoholDTO : BeverageDTO
     {
-        public double Percentage { get; set; }
+        public int Percentage { get; set; }
         public int Age { get; set; }
         
         public AlcoholDTO(int id, byte[] picture, string name, int size, decimal price, int percentage, int age) : base(id, picture, name, size, price)
